coverage: switch from flagSet to global flags
As explained on
https://www.elastic.co/blog/code-coverage-for-your-golang-system-tests :

Before the implementation of the system test coverage we used
FlagSet to keep flags as local as possible. Unfortunately with
the coverage binary these didn't work as expected. For the
test binary, all flags must be known before the execution of
the tests. If a flag is defined during the execution of a
test, the test binary will complain during startup that the
flag is not defined.

Part of https://github.com/rfjakob/gocryptfs/issues/223 .